--- EssentialApp/EssentialApp/ViewController.swift
@@ -1,17 +1,0 @@
-//
-// Created by Rodrigo Porto.
-// Copyright © 2025 PortoCode. All Rights Reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
